+++ microControl/Coord.h
@@ -10,8 +10,8 @@ class Coord {
 public:
 	Coord();
 	Coord(uint8_t x, uint8_t y);
-	uint8_t getX() const;
-	uint8_t getY() const;
+	int getX() const;
+	int getY() const;
 	void setX(uint8_t x);
 	void setY(uint8_t y);
 	bool operator==(const Coord& c);
